Add several utility functions to manage tweets in DB
Add utility functions to help manage the tweets. store_tweets() takes
the tweets from JSON files to a MongoDB. copy_sample() copies a small
sample of the tweets to a second collection, for ease of testing.
drop_duplicates() removes tweets that have the same "id" key.
count_distinct() counts the number of distinct tweets (based on the "id"
key).