Parameterize collection names
Add "collection" arguments to three of the functions to allow the
target collection to be configured from __main__, rather than having
to be changed in each function individually.

The script currently does nothing, as none of these functions should
be executed unknowingly. It can currently (if comments in lines 12x
are removed) can be used to check for duplicate records in
tweets_small and remove extras.